Add files via upload
In this Project, we created an Alarm Clock/Timer where the Alarm can be set up to 1 hour from the point of startup and a clock will keep the current time, you can reset the clock or reset the alarm time chosen which are displayed on 4 7-Segment Displays, the reset switches are placed on switches and other switches help select the minute or seconds digit for the alarm setting and a push button is used to cycle through the digits 0 to 9. A debounce finite state machine circuit is used to ensure the push button is read properly. This project utilizes the song player project where when the alarm goes off it plays a song.